move GStorage to common
